Change: Respect DT on connect
The backend (python) was using the timestamp of the first websocket
connection as the reference timestamp for all further events. If the
connection was established at T, for example, an event with a time delta
of 14 would be inserted with the timestamp T+14.

This behaviour has now changed: the client (javascript) declares its
current delta upon each connect. It can now define its initial delta to
be 20, for example. A following request with a delta of 14 will now be
inserted as T-6.

Jira: MCS-43